requirements.txt
tiktoken
openai
python-dotenv
autogen-ext[azure]
autogen-agentchat
autogen-ext[magentic-one,openai,azure]
olefile 
